Restructure the services folder, adding org related services to a folder, moving a few other files.
