+++ app/app/(dashboard)/customers/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use server";
 
+import { getCustomerOfVendor } from "@/app/services/organization/vendor-organization-service";
 import { getCurrentOrganizationId } from "@/app/services/user-context-service";
-import { getCustomerOfVendor } from "@/app/services/vendor-organization-service";
 import PageHeader from "@/components/common/page-header";
 import ChargeCard from "@/components/customer/charge-card";
 import SubscriptionCard from "@/components/customer/subscription-card";
