+++ app/app/(dashboard)/research/shortlisted/page.tsx
@@ -1,4 +1,4 @@
-import { getShortlistedLeads } from "@/app/services/lead-service";
+import { getShortlistedLeads } from "@/app/services/research/lead-service";
 import { ShortlistedResearch } from "@/components/research/shortlisted-research";
 
 export default async function ShortlistedResearchPage() {
