+++ app/app/(dashboard)/team/team-member-edit-role-btn.tsx
@@ -1,7 +1,10 @@
 "use client";
 
 import { OrganizationRole } from "@/app/generated/prisma";
-import { changeTeamMemberRole, transferOwnership } from "@/app/services/team-management-service";
+import {
+  changeTeamMemberRole,
+  transferOwnership
+} from "@/app/services/organization/team-management-service";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import {
