+++ app/app/(nonav)/join/[inviteId]/accept-invite-btn.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { setCurrentOrganization } from "@/app/services/auth-service";
-import { acceptInvitation } from "@/app/services/team-management-service";
+import { acceptInvitation } from "@/app/services/organization/team-management-service";
 import { Button } from "@/components/ui/button";
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
