+++ components/login/customer-login.tsx
@@ -2,7 +2,7 @@
 
 import useCurrentSession from "@/app/hooks/use-current-session";
 import { userExists } from "@/app/services/auth-service";
-import { setSignUp } from "@/app/services/registration-service";
+import { setSignUp } from "@/app/services/onboarding/registration-service";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
