+++ components/research/search-result.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { addLeadToShortlist } from "@/app/services/lead-service";
+import { addLeadToShortlist } from "@/app/services/research/lead-service";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
