+++ lib/auth.ts
@@ -2,7 +2,7 @@ import Session from "@/app/models/Session";
 import AuthService from "@/app/services/auth-service";
 import { sendVerificationEmail } from "@/app/services/email-service";
 import { defaultOnboardingState } from "@/app/services/onboarding/onboarding-steps";
-import { upsertUser } from "@/app/services/registration-service";
+import { upsertUser } from "@/app/services/onboarding/registration-service";
 import prisma from "@/lib/prisma";
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import { getServerSession, type NextAuthOptions } from "next-auth";
