+++ middleware.ts
@@ -3,13 +3,13 @@ import { withAuth } from "next-auth/middleware";
 import { NextRequest, NextResponse } from "next/server";
 import { OrganizationType } from "./app/generated/prisma";
 import { SessionUser } from "./app/models/Session";
+import { canViewPath, createAuthContext } from "./app/services/organization/role-service";
 import {
   getGhUsernameFromRequest,
   getReservedSubdomainFromRequest,
   getSubdomainFromRequest,
   isVercelPreview
-} from "./app/services/domain-request-service";
-import { canViewPath, createAuthContext } from "./app/services/role-service";
+} from "./app/services/site/domain-request-service";
 
 export const config = {
   matcher: [
